+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.17.5
+
+* added terminate_on application to prepare in TerminableThread